+++ src/Symfony/Component/Notifier/Bridge/Telegram/TelegramTransport.php
@@ -52,17 +52,21 @@ final class TelegramTransport extends AbstractTransport
         private ?string $chatChannel = null,
         ?HttpClientInterface $client = null,
         ?EventDispatcherInterface $dispatcher = null,
+        private readonly bool $disableHttps = false,
     ) {
         parent::__construct($client, $dispatcher);
     }
 
     public function __toString(): string
     {
-        if (null === $this->chatChannel) {
-            return \sprintf('telegram://%s', $this->getEndpoint());
+        $toString = \sprintf('telegram://%s', $this->getEndpoint());
+        $formattedOptions = http_build_query(['channel' => $this->chatChannel, 'disable_https' => $this->disableHttps ?: null]);
+
+        if ($formattedOptions) {
+            $toString .= \sprintf('?%s', $formattedOptions);
         }
 
-        return \sprintf('telegram://%s?channel=%s', $this->getEndpoint(), $this->chatChannel);
+        return $toString;
     }
 
     public function supports(MessageInterface $message): bool
@@ -104,8 +108,9 @@ final class TelegramTransport extends AbstractTransport
         $method = $this->getPath($options);
         $this->ensureExclusiveOptionsNotDuplicated($options);
         $options = $this->expandOptions($options, 'contact', 'location', 'venue');
+        $protocolSchema = $this->disableHttps ? 'http' : 'https';
 
-        $endpoint = \sprintf('https://%s/bot%s/%s', $this->getEndpoint(), $this->token, $method);
+        $endpoint = \sprintf('%s://%s/bot%s/%s', $protocolSchema, $this->getEndpoint(), $this->token, $method);
 
         $response = $this->client->request('POST', $endpoint, [
             $optionsContainer => array_filter($options),